step3 미션 수행
